move vbis tests to dedicated file
