+++ tests/test_inference.py
@@ -259,45 +259,3 @@ def test_owl_inference_tags_reasonable():
     res2 = filter_bnodes(res2)
 
     assert set(res2) == set(map(lambda x: (x,), expected))
-
-
-def test_vbis_to_brick_inference():
-    session = VBISTagInferenceSession()
-    assert session is not None
-
-    # input a fully-qualified VBIS tag, get Brick classes out
-    test_cases = [
-        ("ME-AHU-Su", BRICK.AHU),
-        ("ME-AHU-Su-BU", BRICK.AHU),
-        ("ME-ATU-VAV-SD", BRICK.VAV),
-    ]
-    for (vbistag, brickclass) in test_cases:
-        predicted_classes = session.lookup_brick_class(vbistag)
-        assert brickclass in predicted_classes
-
-
-def test_brick_to_vbis_inference_with_owlrl():
-    ALIGN = Namespace("https://brickschema.org/schema/1.1/Brick/alignments/vbis#")
-
-    # input brick model; instances should have appropriate VBIS tags
-    g = Graph(load_brick=True)
-    data = pkgutil.get_data(__name__, "data/vbis_inference_test.ttl").decode()
-    g.load_file(source=io.StringIO(data))
-    g.expand(profile="owlrl")
-    g.expand(profile="vbis")
-
-    test_cases = [
-        ("http://bldg#f1", "ME-Fa"),
-        ("http://bldg#rtu1", "ME-ACU"),
-    ]
-    for (entity, vbistag) in test_cases:
-        query = f"SELECT ?tag WHERE {{ <{entity}> <{ALIGN.hasVBISTag}> ?tag }}"
-        res = list(g.query(query))
-        assert len(res) == 1
-        assert str(res[0][0]) == vbistag
-
-    conforms, _, results = g.validate()
-    assert conforms, results
-
-
-# TODO: do without owlrl inference
